create method for games ends and update old test with new update
