--- features/airforce_features/movement_plane.feature
@@ -1,8 +1,0 @@
-Feature: All possible movements and rotations of the plane
-
-    @air_force
-    Scenario: Rotation of the plane
-        Given A plane on the map
-        When I make the rotation
-        Then I should obtain a '201' response
-        And The head is in the same place as before
